Issue #108 - tests for expanded date parsing
* Refactor all of the parsing into date parser
* Clean up the uses as much as possible
